requirements.txt
azure-cognitiveservices-speech==1.32.1
ipykernel==6.25.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
moviepy==1.0.3
numpy==1.26.0
pandas==2.1.1
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
streamlit==1.27.0
urllib3==2.0.5
uuid==1.30